"Реализован класс, описывающий блок поиска"
