--- src/main/java/task/pages/ResultPage.java
@@ -1,4 +1,0 @@
-package task.pages;
-
-public class ResultPage extends BasePage{
-}
