+++ src/test/java/Cucumber/CucumberRunner.java
@@ -5,7 +5,7 @@ import io.cucumber.junit.CucumberOptions;
 import org.junit.runner.RunWith;
 
 @RunWith(Cucumber.class)
-@CucumberOptions(features = {"src/test/resources"}, glue = {"steps", "hooks"}, plugin = {"pretty", "html:target/cucumber-html-report",
+@CucumberOptions(features = {"src/test/resources"}, glue = {"steps", "stepsDefinitions"}, plugin = {"pretty", "html:target/cucumber-html-report",
         "json:target/cucumber-report.json"})
 
 public class CucumberRunner {
